auth por atributos sponsor y finder

+++ api/controllers/sponsorshipController.js
@@ -23,16 +23,14 @@ exports.list_my_sponsorships = async function(req, res) {
   console.log("id");
   var idToken = req.headers['idtoken'];//WE NEED the FireBase custom token in the req.header['idToken']... it is created by FireBase!!
   var authenticatedUserId = await authController.getUserId(idToken);
-  if (authenticatedUserId == req.params.sponsor_id){
-    Sponsorship.find(function(err, sponsorships) {
-      if (err){
-        res.status(500).send(err);
-      }
-      else{
-        res.json(sponsorships);
-      }
-    });
-  }
+  Sponsorship.find({sponsor_id : authenticatedUserId}, function(err, sponsorships) {
+    if (err){
+      res.status(500).send(err);
+    }
+    else{
+      res.json(sponsorships);
+    }
+  });
 };
 
 
@@ -69,6 +67,26 @@ exports.create_an_sponsorship = function(req, res) {
 
  exports.read_an_sponsorship = function(req, res) {
   
+  Sponsorship.findById({_id: req.params.sponsorshipId}, async function(err, sponsorship)
+  {
+    if (err){
+      res.send(err);
+    }
+    else{
+      var idToken = req.headers['idtoken'];
+      var authenticatedUserId = await authController.getUserId(idToken);
+      if (String(authenticatedUserId) === String(sponsorship.sponsor_id))
+      {
+        res.json(sponsorship);
+      }
+      else{
+        res.status(403); //Auth error
+        res.send('The Sponsor trying to read is not the owner of this Sponsorship');
+      } 
+    }
+  });
+
+    /*
     Sponsorship.findOne({_id: req.params.sponsorshipId}, async function(err, sponsorship) {
       var idToken = req.headers['idtoken'];//WE NEED the FireBase custom token in the req.header['idToken']... it is created by FireBase!!
       var authenticatedUserId = await authController.getUserId(idToken);
@@ -85,7 +103,7 @@ exports.create_an_sponsorship = function(req, res) {
       res.status(403); //Auth error
       res.send('The Actor is trying to update an Actor that is not himself!');
     }  
-    });
+    });*/
  
 }; 
 
@@ -94,7 +112,17 @@ exports.update_an_sponsorship = function(req, res) {
   //Check if the sponsorship has been previously assigned or not
   //Assign the sponsorship to the proper clerk that is requesting the assigment
   //when updating delivery moment it must be checked the clerk assignment and to check if it is the proper clerk and if not: res.status(403); "an access token is valid, but requires more privileges"
-    Sponsorship.findOneAndUpdate({_id: req.params.sponsorshipId}, req.body, {new: true}, function(err, sponsorship) {
+  Sponsorship.findById({_id: req.params.sponsorshipId}, async function(err, sponsorship)
+  {
+    if (err){
+      res.send(err);
+    }
+    else{
+      var idToken = req.headers['idtoken'];
+      var authenticatedUserId = await authController.getUserId(idToken);
+      if (String(authenticatedUserId) === String(sponsorship.sponsor_id))
+      {
+        Sponsorship.findOneAndUpdate({_id: req.params.sponsorshipId}, req.body, {new: true}, function(err, sponsorship) {
           if (err){
             res.status(500).send(err);
           }
@@ -102,22 +130,50 @@ exports.update_an_sponsorship = function(req, res) {
             res.json(sponsorship);
           }
         });
+      }
+      else{
+        res.status(403); //Auth error
+        res.send('The Sponsor trying to read is not the owner of this Sponsorship');
+      } 
+    }
+  });  
+  
+  
 };
 
 
 exports.delete_an_sponsorship = function(req, res) {
   //Check if the sponsorship were delivered or not and delete it or not accordingly
   //Check if the user is the proper customer that posted the sponsorship and if not: res.status(403); "an access token is valid, but requires more privileges"
-  Sponsorship.deleteOne({
-    _id: req.params.sponsorshipId
-  }, function(err, sponsorship) {
+  
+  Sponsorship.findById({_id: req.params.sponsorshipId}, async function(err, sponsorship)
+  {
     if (err){
-      res.status(500).send(err);
+      res.send(err);
     }
     else{
-      res.json({ message: 'Sponsorship successfully deleted' });
+      var idToken = req.headers['idtoken'];
+      var authenticatedUserId = await authController.getUserId(idToken);
+      if (String(authenticatedUserId) === String(sponsorship.sponsor_id))
+      {
+        Sponsorship.deleteOne({
+          _id: req.params.sponsorshipId
+        }, function(err, sponsorship) {
+          if (err){
+            res.status(500).send(err);
+          }
+          else{
+            res.json({ message: 'Sponsorship successfully deleted' });
+          }
+        });
+      }
+      else{
+        res.status(403); //Auth error
+        res.send('The Sponsor trying to read is not the owner of this Sponsorship');
+      } 
     }
   });
+  
 };
